Fix fallback profile image path
